fix: convert amount to float

+++ kartoza/custom_js/employee_benefit_claim.js
@@ -15,13 +15,13 @@ frappe.ui.form.on('Employee Benefit Claim', {
 			async:false,
 			callback: function(r){
 					if(!r.exc){
-						r.message["amount_per_kilometer"] = parseInt(r.message["amount_per_kilometer"])
+						r.message["amount_per_kilometer"] = parseFloat(r.message["amount_per_kilometer"])
 						if(!r.message["amount_per_kilometer"]){
 							frappe.throw("Set Amount Per Kilometer in HR Settings")
 						}
 						frm.set_value("claimed_amount",frm.doc.kilometer * r.message["amount_per_kilometer"]);
 					}
-			} 
+			}
 		})
 	}
 });
